Implemented the Binary_code class; some bug fixes.
A Binary_code instance is a set of binary codewords of specified length,
which can be given as a list, read from a file, or generated using a
generator matrix (passed directly or read from a file).
Implemented all the initiation/reading/adding methods, simple things
like size and __str__.
Re-wrote the previously written fuctions for finding the min and max
Hamming distance and for finding the bit sum distribution (i.e.  how
many codewords have what sum of bits) as class methods.
TODO: Now need to implement manipulation, such as adding a parity bit or
filtering by the sum or Hamming distance (added empty prototypes for
those).

Added __xor__ etc to Binary_codeword so that it actually implements
the bitwise operators I need!  (just let them act on the underlying
bitstrings)

Tested everything on some simple codes and the [19,10,5] code from a
generator matrix.  It all seems to work.