Implemented an invert() method for Binary_code.
Returns new code with all codewords inverted bitwise.
The minimum Hamming distance is conserved, as is the code length and
size. What changes is the number of codewords with a given weight.

The main point of this is that we can go from a code that contains the
all-zero word (which the standard-form codes always do) to one that
doesn't, since for our deepseq combinatorial pooling application the
all-zero word is unusable because that pool would never be sequenced.

Testing:
>>> B16_ = B16.invert()
>>> B16.find_bit_sum_counts()
[(0, 1), (6, 44), (8, 45), (10, 28), (12, 10)]
>>> B16_.find_bit_sum_counts()
[(4, 10), (6, 28), (8, 45), (10, 44), (16, 1)]
>>> B16.find_Hamming_distance_range()
(6, 12)
>>> B16_.find_Hamming_distance_range()
(6, 12)