Added remove(value) and remove_all_zero_codeword() methods to Binary_code.
The remove(value) method removes Binary_codeword(value) from the
codeword set; raises an exception if it's not there (same as usually
with sets).
The remove_all_zero_codeword() method removes the appropriate-length
all-zero codeword and returns 1 if that codeword was in the set, or
returns 0 if it wasn't (TODO is that the best way to implement that?).

I also added unit tests for both.