Moved the full main functionality into a function; added multiple test cases.
Made a new "run_main_function(parser,options,args)" function that
basically performs the full run.  This makes it easier to run multiple
test cases in a single run.

Made a list with test cases: one corresponding to the 63-sample
plate-pool small-scale run with a single Biomek file, one for the
384-sample well-pool run with multiple Biomek files.  They both run.