Started implementing mirrored pooling (wrote basic structure and notes).
In addition to the usual pooling, we also want a "mirrored" pooling
option:  where each sample is in each pool it WASN'T in in the original
pooling, and vice versa.  Added an option (-M) for that; generated
mirror outfile names in generate_outfile_names and passed them around as
appropriate; wrote the basic setup for doing the mirroring (need a
mirror_transfer_file_command_list function to generate the commands,
then use the normal write_data_to_Biomek_files function to write them to
the set of mirror outfiles) plus some implementation notes in comments.

Also renamed file-writing functions from print_* to write_*.