Implemented giving different names to mirror pool plates; cosmetic changes.
Added an option (-u or --mirror_pool_plate_suffix) for the string to be
added to mirror pool (destination) plate names, if desired.  Implemented
the functionality by a simple renaming on output_plate_names and then
re-generating pool_positions based on that, right in run_main_function.
Added that to the test case (test2), it works.

Renamed some functions/variables:
 - codeword_and_position_lists_to_Biomek_file to make_Biomek_file_commands
 - transfer_file_command_list to Biomek_file_commands

TODO: write unit tests for several of the new functions!