Moved -T test outputs into a directory; other small -T fixes.
Made all -T test output files go into a test_outputs directory instead
of polluting the current directory;  test_outputs is created if it
doesn't exist, or removed if it exists (with a message stating which).

Added a check for the presence of an error-correcting_codes folder,
since the -T tests won't work without it.

Restructured the whole section of the main program related to -T a bit.

Also changed the default --mirror_pool_plate_suffix from empty to
_mirror, since having it empty is stupid (and it made the code shorter),
and removed a related input-check (since it makes no sense now).