Progress on making clonality_count_conflicts faster - doesn't quite work.
Reimplemented/simplified clonality_count_conflicts using an
expand_by_all_mutations function that will take a set of codewords and
return a set of all codewords that are too close to them to be legal
clonality results, given the specified # of allowed changes.

Implemented expand_by_all_mutations, using a help function
_change_positions_to_value (TODO - both need unit testing.)

HOWEVER - as currently done, clonality_count_conflicts won't work right,
because when A+B=X is counted as a clonality conflict because X is close
to C, the codeword that will be given a conflict count for this (in
addition to A and B) will be X, not C, which is bad - X isn't even part
of the code.  So I need to rewrite things, see TODO comment.
This probably fails unit tests, I haven't actually tries.