Added easy method to remove all-one codeword from a code.
Renamed Binary_code remove_all_zero_codeword function to
remove_extreme_codeword and rewrote it to allow removing either the
all-zero codeword (default) or the all-one codeword (which is bad for
clonality-conflict reasons).
Changed the rest of the program to use the new function properly.
Added unit-test case for the new functionality. All tests pass.