Moved run-test to own function; small bug-fix.
Moved the run-test section of __main___ to a separate function,
do_test_run.  (This setup is based on ../../../mutant_pool_screens/
mutant_deepseq_analysis/code/mutant_count_alignments.py)

Uncovered a bug while doing this - get_binary_code used
options.number_of_pools but wasn't passed options as an argument, it was
just treating it as global. That's bad. Fixed by giving it a length
argument and calling it with the value of that argument based on
options.number_of_pools.

Improved an error message and except condition in
binary_code_utilities.py based on the bug results.