Fixed error-correcting code names (3-3-1, not 3-2-1, and 4-3-2, not 4-2-2).
I was using wrong names for the two error-correcting codes I was using -
obviously if d is 1, n and k should be the same, or else the code is
just throwing away information for no reason.

I also renamed the two relevant files in error-correcting_codes folder
(which is not tracked by git).