Added custom plate types (incl. fake6); related changes, tests.
Modified the Plate_type class to allow custom plate definitions (by
passing in the sequential list of well IDs directly) as well as
pre-defined standard plate shapes:
 - rewrote __init__ to take an optional well_ID_list argument and use it
   if provided (otherwise use the size argument as before)
 - renamed make_well_ID_list_and_dict to make_well_ID_list_from_size,
   and moved the well_ID_dict generation out of it into __init__, since
   standard plates need that function to make their well_ID_list, but
   custom plates don't (since they get well_ID_list passed as an
   argument), but both need well_ID_dict generated
 - moved some of the size/row/column related code from __init__ to
   make_well_ID_list_from_size, since there's no need for self.rows and
   self.columns to exist and be visible from the outside, and custom
   plates may not even HAVE consistent rows and columns.

Modified plate_types to also include a custom "fake6" plate type
(well_ID_list=['B2','B7','B11','G2','G7','G11'] - the wells are defined
as for a 96-well plate but in the positions where 6-well wells are, so
we can put a physical 6-well plate there and pretend to the Biomek that
it's a 96-well plate and it'll work, since the Biomek seems to have
trouble accepting a 6-well plate specification).
Also renamed plate_types to plate_type_definitions for clarity, and
added string as well as integer keys for the standard plate sizes for
ease of use.
In Plate_type class, renamed plate_types_rows_columns to
standard_plate_shapes for clarity, and removed plate_sizes because
there's no reason for it - added a module-level defined_plate_types list
instead, which is just a reduced and sorted list of
plate_type_definitions keys.  Went over all of the code to make sure
either defined_plate_types or plate_type_definitions.keys() was used in
place of plate_sizes, whichever was appropriate (generally
plate_type_definitions.keys() for testing but defined_plate_types for
printing).

Rewrote numbers_to_plate_and_well_IDs a bit to match the new code -
mostly to deal with the fact that plate_type_name is no longer
necessarily an integer equal to plate_size (as in the fake6 case).  Did
a lot of renaming variables for clarity.

In define_option_parser, changed the source/destination plate size
options (-s and -S) from an int type to a choice type based on
defined_plate_types.

Modified the unit-tests:
 - renamed plate_types to plate_type_definitions etc
 - added the 'fake6' plate type to all tests
 - added test__creating_good_plate_types_from_list and
   test__creating_bad_plate_types_from_list test functions to test the
   new custom plate functionality

Added a new run-test for the fake6 plate type (test_fake6-plate - same
as test_basic but using fake6 destination plate)

All unit-tests and run-tests pass.