Switched run-tests to match new codeword-subset choice method.
Modified all run-test output reference files by hand to match the
changes described in the last commit.  The only actual change that was
relevant to the run-tests was that the codeword sort is now
purely lexicographic (by binary string) instead of by bit-sum first, so
011 sorts before 100 instead of after, so I switched the orders of a few
lines and the corresponding sequential sample numbers and positions, and
the order of the bits in the pool pooling_schemes.

All run-tests pass (all the unit-tests do too).