IN PROGRESS: adding well-cycling plate types - STILL DOESN'T WORK.
Rewrote numbers_to_plate_and_well_IDs to return a list of generators
instead of a list of positions (and also a list of all-position strings
to print in the general info output file),
Modified the unit-tests and the rest of the code to match that, but it
DOESN'T WORK, because I'm trying to use generators as functions and
they're not.  Also doing the tests like this is hard and strange.

Added Plate_type.get_all_well_IDs_from_number method to print ALL the
well IDs for a given position in the general info output file.

DOESN'T WORK - unit-tests fail, run-tests and even smoke-tests fail.