Adjusted for changed bitstring module behavior; tests pass (with new module).
Simple change: bitstring.BitArray(bin='111').bin used to be 0b111, now
is just 111!  I used to strip the first 2 characters to compensate, so
with the new version all my codewords ended up too short by 2.
Added a test for either behavior after bitstring import; defined two
functions to get a binary string from a bitstring, and the
Binary_codeword.string function is set to the right one at class
definition, to avoid slowdown.

All tests pass with the new bitstring version; I haven't hunted down an
old version of the bitstring module to test if this still works with it,
but it should, it's not exactly complicated.